requirements.txt
langchain-google-genai
google-generativeai
langchain
streamlit
python-dotenv
PyPDF2

